Rename Random() to Prng()
