+++ tools/acornui-utils/js/src/main/kotlin/com/acornui/math/Random.kt
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+@file:Suppress("DEPRECATION")
+
 package com.acornui.math
 
 import kotlin.js.Math
